Homework1: Sergei Krivchik
