+++ src/Homework1/1.py
@@ -1,0 +1,1 @@
+print ('Hello, world!')
