+++ src/Homework1/14.py
@@ -1,0 +1,14 @@
+def median(pool):
+    '''Statistical median to demonstrate doctest.
+    >>> median([2, 9, 9, 7, 9, 2, 4, 5, 8])
+    6 #change to 7 in order to pass the test
+    '''
+    copy = sorted(pool)
+    size = len(copy)
+    if size % 2 == 1:
+        return copy[int((size - 1) / 2)]
+    else:
+        return (copy[int(size/2 - 1)] + copy[int(size/2)]) / 2
+if __name__ == '__main__':
+    import doctest
+    doctest.testmod()
