+++ src/Homework1/2.py
@@ -1,0 +1,2 @@
+name = input('What is your name?\n')
+print ('Hi, %s.' % name)
