+++ src/Homework1/3.py
@@ -1,0 +1,3 @@
+friends = ['john', 'pat', 'gary', 'michael']
+for i, name in enumerate(friends):
+    print ("iteration {iteration} is {name}".format(iteration=i, name=name))
