+++ src/Homework1/4.py
@@ -1,0 +1,4 @@
+parents, babies = (1, 1)
+while babies < 100:
+    print ('This generation has {0} babies'.format(babies))
+    parents, babies = (babies, parents + babies)
