+++ src/Homework1/5.py
@@ -1,0 +1,6 @@
+def greet(name):
+    print ('Hello', name)
+
+greet('Jack')
+greet('Jill')
+greet('Bob')
