+++ src/Homework1/stocks.csv
@@ -1,0 +1,3 @@
+GOOG,"Google, Inc.",505.24,0.47,0.09
+YHOO,Yahoo! Inc.,27.38,0.33,1.22
+CNET,"CNET Networks, Inc.",8.62,-0.13,-1.4901
